+++ print_u.c
@@ -6,15 +6,15 @@
 /*   By: ade-la-c <ade-la-c@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/08 19:00:41 by ade-la-c          #+#    #+#             */
-/*   Updated: 2020/10/09 14:42:33 by ade-la-c         ###   ########.fr       */
+/*   Updated: 2020/10/16 16:29:28 by ade-la-c         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-static long				printnbr(t_flag f, long num, int tag)
+static long long		printnbr(t_flag f, long num, int tag)
 {
-	long					i;
+	long				i;
 
 	i = 0;
 	if (!f.prec && num == 0)
@@ -32,9 +32,9 @@ static long				printnbr(t_flag f, long num, int tag)
 
 void					ft_print_u(t_flag f, va_list args)
 {
-	long					num;
+	long long			num;
 
-	num = va_arg(args, long);
+	num = va_arg(args, long long);
 	if (f.zero && f.width && f.prec == -1)
 	{
 		f.prec = (num < 0 ? f.width - 1 : f.width);
